refactor(editors): drop MDXEditor — every markdown surface is now Lexical
Replace the MDXEditor wrapper (`SpeakerLetterEditor` + `EditorSection`)
with a canonical Lexical `MarkdownEditor` at
`src/features/templates/MarkdownEditor.tsx`. Same drop-in API
(`initialMarkdown`, `onChange`, `ariaLabel`) plus an optional
`variables` / `groupLabels` pair that turns on the Insert-Variable
dropdown when the host needs one. `{{token}}` text round-trips
through inline visual chips on the way in and back out, so the
schema and renderers (LetterCanvas, send-time interpolation, message
template previews) keep working unchanged.

Updated consumers — all now import from `MarkdownEditor`:

- MessageTemplateCardDesktop (six message templates)
- TemplateEditorModal (mobile fullscreen template editor)
- PrepareInvitationLetterTab (per-speaker letter override)
- InviteMessageOverridePanel (Invite Member greeting override)
- SpeakerLetterEditorColumn (ward-level speaker letter editor)

Toolbar: `ProgramToolbar`'s `variables` / `groupLabels` props are
optional now so the Insert-Variable menu hides cleanly when the
host has nothing to insert.

Cleanup: deleted the old MDXEditor wrapper file and the now-redundant
`SpeakerLetterMarkdownEditor` stub. `pnpm remove @mdxeditor/editor`
removes the package and 188 transitive deps from the lockfile.
