+++ src/features/templates/MarkdownEditor.tsx
@@ -16,35 +16,51 @@ import {
 } from "@lexical/markdown";
 import { lexicalTheme } from "@/features/program-templates/lexicalTheme";
 import { ProgramToolbar } from "@/features/program-templates/ProgramToolbar";
+import type { ProgramVariable } from "@/features/program-templates/programVariables";
 import { VariableChipNode } from "@/features/program-templates/nodes/VariableChipNode";
 import { VARIABLE_CHIP_MARKDOWN_TRANSFORMER } from "@/features/program-templates/nodes/variableChipMarkdown";
-import { SPEAKER_LETTER_GROUP_LABEL, SPEAKER_LETTER_VARIABLES } from "./speakerLetterVariables";
 
 interface Props {
-  /** Markdown source to seed the editor with. Round-trips through
-   *  `$convertFromMarkdownString` / `$convertToMarkdownString` so the
-   *  rest of the speaker-letter pipeline (LetterCanvas, send-time
-   *  interpolation) sees plain markdown unchanged. */
   initialMarkdown: string;
   onChange: (markdown: string) => void;
   ariaLabel: string;
+  /** Optional Insert-variable surface. Omit both to hide the menu —
+   *  every editor still round-trips `{{token}}` text into chips for
+   *  authoring even when no menu is shown. */
+  variables?: readonly ProgramVariable[];
+  groupLabels?: Readonly<Record<string, string>>;
+  /** Placeholder copy when the editor is empty. */
+  placeholder?: string;
 }
 
 const NODES = [HeadingNode, QuoteNode, ListNode, ListItemNode, VariableChipNode];
 const ALL_TRANSFORMERS: Transformer[] = [VARIABLE_CHIP_MARKDOWN_TRANSFORMER, ...TRANSFORMERS];
 
-/** Lexical editor for the speaker-letter body. Markdown in / markdown
- *  out — the chip transformer converts `{{token}}` to inline visual
- *  chips for authoring while keeping the bytes on disk plain markdown
- *  so the existing letter renderer (`LetterCanvas`) keeps working
- *  with no schema change. */
-export function SpeakerLetterMarkdownEditor({ initialMarkdown, onChange, ariaLabel }: Props) {
+/** Canonical markdown editor for the app — Lexical inside, plain
+ *  markdown on the wire. `{{token}}` text round-trips through inline
+ *  visual chips so the existing markdown-based render pipeline keeps
+ *  working while the bishop sees a structured editor. Drop-in for
+ *  every place we previously used MDXEditor.
+ *
+ *  Layout: rounded chalk wrapper + sticky walnut toolbar + scrollable
+ *  content. Wrapper claims its parent's height when the parent is
+ *  height-constrained (`lg:h-full lg:min-h-0` chain); otherwise the
+ *  scroll container's `min-h-70` keeps the editor a usable 280 px
+ *  in natural-flow contexts (modals, settings sections). */
+export function MarkdownEditor({
+  initialMarkdown,
+  onChange,
+  ariaLabel,
+  variables,
+  groupLabels,
+  placeholder = "Write here…",
+}: Props) {
   const initialConfig = {
-    namespace: "SpeakerLetterEditor",
+    namespace: "MarkdownEditor",
     theme: lexicalTheme,
     nodes: NODES,
     onError: (e: Error) => {
-      console.error("[SpeakerLetterMarkdownEditor]", e);
+      console.error("[MarkdownEditor]", e);
       throw e;
     },
     editorState: () => $convertFromMarkdownString(initialMarkdown, ALL_TRANSFORMERS),
@@ -54,12 +70,9 @@ export function SpeakerLetterMarkdownEditor({ initialMarkdown, onChange, ariaLab
     <div className="rounded-lg border border-border bg-chalk overflow-hidden flex flex-col lg:h-full lg:min-h-0">
       <LexicalComposer initialConfig={initialConfig}>
         <div className="sticky top-0 z-10 bg-parchment-2 shadow-[0_1px_0_var(--color-border)]">
-          <ProgramToolbar
-            variables={SPEAKER_LETTER_VARIABLES}
-            groupLabels={SPEAKER_LETTER_GROUP_LABEL}
-          />
+          <ProgramToolbar variables={variables} groupLabels={groupLabels} />
         </div>
-        <div className="relative flex-1 min-h-0 overflow-y-auto">
+        <div className="relative flex-1 min-h-70 overflow-y-auto">
           <RichTextPlugin
             contentEditable={
               <ContentEditable
@@ -69,7 +82,7 @@ export function SpeakerLetterMarkdownEditor({ initialMarkdown, onChange, ariaLab
             }
             placeholder={
               <div className="pointer-events-none absolute top-3 left-4 font-serif italic text-[14px] text-walnut-3">
-                Write the letter — text, lists, scripture, variables…
+                {placeholder}
               </div>
             }
             ErrorBoundary={LexicalErrorBoundary}
@@ -79,6 +92,27 @@ export function SpeakerLetterMarkdownEditor({ initialMarkdown, onChange, ariaLab
         <ListPlugin />
         <MarkdownOnChange onChange={onChange} />
       </LexicalComposer>
+    </div>
+  );
+}
+
+interface SectionProps extends Omit<Props, "ariaLabel"> {
+  label: string;
+  /** Optional override for the editor's aria-label. Defaults to `label`. */
+  ariaLabel?: string;
+  disabled?: boolean;
+}
+
+/** Labelled wrapper around `MarkdownEditor` matching the previous
+ *  `EditorSection` API consumed by the message template cards,
+ *  template editor modal, and per-speaker letter override panel. */
+export function EditorSection({ label, disabled, ...rest }: SectionProps) {
+  return (
+    <div className={disabled ? "opacity-60 pointer-events-none" : ""}>
+      <div className="font-mono text-[10px] uppercase tracking-[0.14em] text-walnut-3 mb-1.5">
+        {label}
+      </div>
+      <MarkdownEditor {...rest} ariaLabel={rest.ariaLabel || label} />
     </div>
   );
 }
